pyproject.toml
[tool.poetry]
name = "osf-bot"
version = "0.1.0"
description = "A community bot used in the Manipal OSF server."
authors = ["dhzdhd"]

[tool.poetry.dependencies]
python = "^3.9"
disnake = "^2.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-annotations = "~=2.0"
flake8-bugbear = "^22.1.11"
flake8-docstrings = "~=1.4"
flake8-import-order = "~=0.18"
flake8-string-format = "~=0.2"
flake8-tidy-imports = "~=4.0"
flake8-todo = "~=0.7"
pep8-naming = "~=0.9"
taskipy = "^1.9.0"
pre-commit = "^2.15.0"
python-dotenv = "^0.19.1"
black = "^22.1.0"

[tool.taskipy.tasks]
bot = { cmd = "python -m osfbot", help = "Runs Bot"}
lint = { cmd = "pre-commit run --all-files", help = "Lints project" }
precommit = { cmd = "pre-commit install", help = "Installs the pre-commit git hook" }
format = { cmd = "black osfbot", help = "Runs the black python formatter" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
